+++ waku/v2/utils/pagination.nim
@@ -10,5 +10,5 @@ type
   Index* = object
     ## This type contains the  description of an Index used in the pagination of WakuMessages
     digest*: MDigest[256]
-    receiverTime*: float64
-    senderTime*: float64 # the time at which the message is generated
+    receiverTime*: int64
+    senderTime*: int64 # the time at which the message is generated
